#629: Leverage BigDecimal classes for decimals calculations

+++ src/main/java/org/benetech/servicenet/matching/counter/UrlSimilarityCounter.java
@@ -1,5 +1,6 @@
 package org.benetech.servicenet.matching.counter;
 
+import java.math.BigDecimal;
 import org.benetech.servicenet.util.UrlNormalizationUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.benetech.servicenet.matching.model.MatchingContext;
@@ -12,10 +13,10 @@ import java.util.Locale;
 public class UrlSimilarityCounter extends AbstractSimilarityCounter<String> {
 
     @Value("${similarity-ratio.weight.url.equal-upper-cased}")
-    private float uppercasedWeight;
+    private BigDecimal uppercasedWeight;
 
     @Override
-    public float countSimilarityRatio(String url1, String url2, MatchingContext context) {
+    public BigDecimal countSimilarityRatio(String url1, String url2, MatchingContext context) {
         if (StringUtils.isBlank(url1) || StringUtils.isBlank(url2)) {
             return NO_MATCH_RATIO;
         }
